Add/change some redirects, change a variable name
Changed status code of redirect in loginHandler to statusFound. Added
redirect in homeHandler when an invalid URL is input. Also changed the
name of the variable hashedPasswordBytes to storedPasswordBytes to reduce
ambiguity. Added an error message that compares the input password to
the stored password when the user inputs the wrong password.